Added auto install bash script (ubuntu/debian) (#155)
I wrote this bash script to automate the process of compiling and install this driver on Ubuntu Xenial, but the script is usable as is (or with minor tweaks) on pretty much any Debian based distro. I thought it might be a useful tool, especially for novice linux users, either as a no-brain install method, or as a guide to demonstrate the steps for installing the driver. I added some comments to explain what each group of commands does.

I realize this might be just be clutter for the project, but I thought I would push it and let you decide if it would be useful or not. Thanks for all the hard work on this driver, it is much appreciated!